Fix all clippy warnings and remove dead code across engine crates
Remove unused wasm_api items (with_state_mut, StratagemView,
decision_surface_to_view, command_to_action_view) that were
superseded by the macro_action conversion path. Fix clippy lints:
iterator patterns over manual indexing in eval_nnue, .clamp() calls,
Box large enum variant, Default impls for search AI types, const
thread_local initializers, struct initializer syntax, redundant
closures, doc indentation, and hex literal grouping.

+++ engine/crates/eval_features/src/lib.rs
@@ -639,7 +639,7 @@ fn extract_global_features(
         battle_round: round_num,
         phase: phase_to_u8(state.current_phase),
         vp_differential: own_player.vp.value() - enemy_player.vp.value(),
-        cp_differential: (own_player.cp.value() - enemy_player.cp.value()) as i8,
+        cp_differential: (own_player.cp.value() - enemy_player.cp.value()),
         own_models_remaining,
         enemy_models_remaining,
         own_model_ratio,
